+++ test/managers/select_manager_test.go
@@ -26,6 +26,9 @@ func TestSelectManager(t *testing.T) {
   _ = mgr.Order(1)
   _ = mgr.Group(1)
   _ = mgr.Having(1)
+  _ = mgr.Union(managers.Selection(relation))
+  _ = mgr.Intersect(managers.Selection(relation))
+  _ = mgr.Except(managers.Selection(relation))
   _ = mgr.Engine(1)
   _, _ = mgr.ToSql()
 }