+++ tree/managers/select_manager.go
@@ -111,6 +111,27 @@ func (self *SelectManager) Having(expr interface{}) *SelectManager {
   return self
 }
 
+// Union sets the SelectManager's Tree's Combination member to a
+// UnionNode of itself and the parameter `manager`'s Tree.
+func (self *SelectManager) Union(manager *SelectManager) *SelectManager {
+  self.Tree.Combinator = nodes.Union(self.Tree, manager.Tree)
+  return self
+}
+
+// Intersect sets the SelectManager's Tree's Combination member to a
+// IntersectNode of itself and the parameter `manager`'s Tree.
+func (self *SelectManager) Intersect(manager *SelectManager) *SelectManager {
+  self.Tree.Combinator = nodes.Intersect(self.Tree, manager.Tree)
+  return self
+}
+
+// Except sets the SelectManager's Tree's Combination member to a
+// ExceptNode of itself and the parameter `manager`'s Tree.
+func (self *SelectManager) Except(manager *SelectManager) *SelectManager {
+  self.Tree.Combinator = nodes.Except(self.Tree, manager.Tree)
+  return self
+}
+
 // Sets the SQL Enginge.
 func (self *SelectManager) Engine(engine interface{}) *SelectManager {
   if _, ok := VISITORS[engine]; ok {